requirements.txt
Django==3.2.13
django-crispy-forms==1.14.0
django-extra-views==0.14.0
Pillow==9.1.1
